requirements.txt
PyYAML==5.4.1
python-gitlab==2.6.0
